Upgrade render system for SDK 4.3: shaders, HAL, filters, mesh, text
Update all renderer backends (D3D9, D3D11, GL) and core render
infrastructure for 4.3 compatibility. Includes regenerated shader
descriptors, updated filter pipeline, text rendering, and mesh cache.
Adds intermediate shader sources for D3D9, GL, and PS4 backends.

+++ Src/Render/FontProvider/Render_FT2Helper.h
@@ -81,12 +81,12 @@ private:
     {
         return float((int)v >> 6);
     }
-    static inline Render::CoordType SF_STDCALL FtToTwips(int v)
+    static inline Render::CoordType SF_STDCALL FtToTwips(long v)
     {
         return float((v * 20) >> 6);
     }
 
-    static inline Render::CoordType SF_STDCALL FtToS1024(int v)
+    static inline Render::CoordType SF_STDCALL FtToS1024(long v)
     {
         return float(v >> 6);
     }
